refactor advisory workspace handoff route
