+++ src/app/router_registry.py
@@ -26,6 +26,9 @@ from app.routers.advisory_policy_evaluations import (
 from app.routers.advisory_workspace_assistant import (
     router as advisory_workspace_assistant_router,
 )
+from app.routers.advisory_workspace_handoff import (
+    router as advisory_workspace_handoff_router,
+)
 from app.routers.advisory_workspace_version_lookups import (
     router as advisory_workspace_version_lookups_router,
 )
@@ -244,6 +247,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(advisory_workspace_versions_router)
     app.include_router(advisory_workspace_version_lookups_router)
     app.include_router(advisory_workspace_assistant_router)
+    app.include_router(advisory_workspace_handoff_router)
     app.include_router(advisory_policy_router)
     app.include_router(advisory_policy_evaluations_router)
     app.include_router(advisory_policy_evaluation_evidence_router)
